-Renamed toJson to toJSON for convention -Moved column mappings to new file -Fixed validation options
